requirements.txt
# Core dependencies
streamlit>=1.32.0,<2.0.0
google-generativeai>=0.4.1,<1.0.0
python-dotenv>=1.0.0,<2.0.0
langchain>=0.1.16,<0.2.0
langchain-google-genai>=0.0.11,<0.1.0

# Development dependencies
black>=23.0.0
flake8>=6.0.0
pytest>=7.0.0
pytest-cov>=4.0.0

# Utility packages
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0

# Documentation
mkdocs>=1.4.0
mkdocs-material>=9.0.0